Refactor: move ccConfig into separate file
