+++ compiler/src/dotty/tools/dotc/cc/CaptureOps.scala
@@ -7,12 +7,10 @@ import Types.*, Symbols.*, Contexts.*, Annotations.*, Flags.*
 import Names.TermName
 import ast.{tpd, untpd}
 import Decorators.*, NameOps.*
-import config.SourceVersion
 import config.Printers.capt
 import util.Property.Key
 import tpd.*
 import StdNames.nme
-import config.Feature
 import collection.mutable
 import CCState.*
 import reporting.Message
@@ -23,57 +21,6 @@ private val Captures: Key[CaptureSet] = Key()
 
 /** Context property to print root.Fresh(...) as "fresh" instead of "cap" */
 val PrintFresh: Key[Unit] = Key()
-
-object ccConfig:
-
-  /** If enabled, use a special path in recheckClosure for closures
-   *  to compare the result tpt of the anonymous functon with the expected
-   *  result type. This can narrow the scope of error messages.
-   */
-  inline val preTypeClosureResults = false
-
-  /** If this and `preTypeClosureResults` are both enabled, disable `preTypeClosureResults`
-   *  for eta expansions. This can improve some error messages.
-   */
-  inline val handleEtaExpansionsSpecially = true
-
-  /** Don't require @use for reach capabilities that are accessed
-   *  only in a nested closure. This is unsound without additional
-   *  mitigation measures, as shown by unsound-reach-5.scala.
-   */
-  inline val deferredReaches = false
-
-  /** Check that if a type map (which is not a BiTypeMap) maps initial capture
-   *  set variable elements to themselves it will not map any elements added in
-   *  the future to something else. That is, we can safely use a capture set
-   *  variable itself as the image under the map. By default this is off since it
-   *  is a bit expensive to check.
-   */
-  inline val checkSkippedMaps = false
-
-  /** Always repeat a capture checking run at least once if there are no errors
-   *  yet. Used for stress-testing the logic for when a new capture checking run needs
-   *  to be scheduled because a provisionally solved capture set was later extended.
-   *  So far this happens only in very few tests. With the flag on, the logic is
-   *  tested for all tests except neg tests.
-   */
-  inline val alwaysRepeatRun = false
-
-  /** After capture checking, check that no capture set contains ParamRefs that are outside
-   *  its scope. This used to occur and was fixed by healTypeParam. It should no longer
-   *  occur now.
-   */
-  inline val postCheckCapturesets = false
-
-  /** If true, turn on separation checking */
-  def useSepChecks(using Context): Boolean =
-    Feature.sourceVersion.stable.isAtLeast(SourceVersion.`3.7`)
-
-  /** Not used currently. Handy for trying out new features */
-  def newScheme(using Context): Boolean =
-    Feature.sourceVersion.stable.isAtLeast(SourceVersion.`3.8`)
-
-end ccConfig
 
 /** Are we at checkCaptures phase? */
 def isCaptureChecking(using Context): Boolean =
